Crate pallet_xcm_helper

Common XCM Helper pallet

Overview

This pallet should be in charge of everything XCM related including callbacks and sending XCM calls.

Re-exports

pub use pallet::*;

Modules

pallet

The module that hosts all the FRAME types needed to add this pallet to a runtime.

weights

Autogenerated weights for pallet_xcm_helper

Traits

XcmHelper

Type Definitions

AccountIdOf

AssetIdOf

BalanceOf

CallIdOf